todo: create utils functions
