+++ libs/todo/src/lib/model/todo.model.ts
@@ -2,21 +2,8 @@ export enum TodoStatus {
   TODO = 'TODO',
   WIP = 'WIP',
   REVIEW = 'REVIEW',
-  COMPLETE = 'COMPLETE',
+  DONE = 'DONE',
 }
-
-export const getNextStatus = (status: TodoStatus): TodoStatus => {
-  switch (status) {
-    case TodoStatus.TODO:
-      return TodoStatus.WIP;
-    case TodoStatus.WIP:
-      return TodoStatus.REVIEW;
-    case TodoStatus.REVIEW:
-      return TodoStatus.COMPLETE;
-    default:
-      return status;
-  }
-};
 
 export interface Todo {
   id: string;
